+++ db_migrations/0.0.0-to-0.8.0.sql
@@ -1,0 +1,3 @@
+ALTER TABLE project ADD COLUMN working_agreement TEXT NOT NULL DEFAULT '';
+ALTER TABLE project ADD COLUMN definition_of_done TEXT NOT NULL DEFAULT '';
+ALTER TABLE project ADD COLUMN continuous_integration_url CHARACTER VARYING NOT NULL DEFAULT '';