+++ CHANGES.md
@@ -3,3 +3,4 @@
 
 - expected monthly worked hours in /times/report/pivot fix
 - current user always should be at the top of list in /times/report/pivot
+- add Definition of Done, Working agreement and Continuous Integration Url to project.